+++ packages/aesthetic-adapter-glamor/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.2.0 - 05/30/18
+#### 🚀 New
+* Updated `aesthetic` peer requirement to 2.5.
+
+#### 🛠 Internal
+* Removed `lodash.merge` dependency.
+
 # 2.1.2 - 03/20/18
 #### 🐞 Fixed
 * Fixed incorrectly built and published files.